extract DemoProviderTooltip into its own component
